+++ RELEASE_NOTES.md
@@ -1,6 +1,23 @@
+### Version 1.4.0
+- added method (no app) KButil_update_genome_lineage_from_files()
+
+### Version 1.3.0
+- added method and app to copy an object of a limited number of common types
+- fixed bug in KButil_Concat_MSAs() where f_id was not split from genome_id
+
+### Version 1.2.0
+- added method and app to bulk hide objects by type from workspace
+- added method and app to bulk unhide objects by type from workspace
+
+### Version 1.1.0
+- added method and app KButil_count_ws_objects()
+- added method and app KButil_update_genome_species_name()
+- added method (no app) KButil_update_genome_fields_from_files()
+- added method (no app) KButil_update_genome_features_from_file()
+
 ### Version 1.0.0
 - Initial release version
 
 ### Version 0.0.1
 - Initial dev version
-
+- added method and app KButil_Concat_MSAs()